+++ CHANGELOG.md
@@ -37,6 +37,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/756>)
 - Add jupyter notebook example rendering to documentation
   (<https://github.com/openvinotoolkit/datumaro/pull/758>)
+- Add an interface to manipulate 'infos' to store the dataset meta-info
+  (<https://github.com/openvinotoolkit/datumaro/pull/767>)
 
 ### Changed
 - Get list of subsets and support only Image media type in visualizer